+++ src/python/move_emojis/.gitignore
@@ -1,4 +1,4 @@
-move_consts.txt
+data/move_consts.txt
 data/base_emojis.json
 data/viable_emojis.json
 data/zwj_emojis.json
